Backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework>=3.9.0,<3.13.1
django-cors-headers==3.11.0
django-rest-auth==0.3.3
django-allauth ==0.50.0